KEY-000 Use fiat-crypto wherever possible. (#325)
* KEY-000 Use fiat-crypto wherever possible.

* lint

* KEY-000 Remove garbage.

+++ pkg/base/curves/bls12381/scalar_field.go
@@ -249,7 +249,7 @@ func (sf *ScalarField[_]) Select(choice bool, x0, x1 curves.Scalar) curves.Scala
 	if !ok0 || !ok1 || oks {
 		panic("Not a BLS12381 scalar")
 	}
-	s.V.CMove(x0s.V, x1s.V, utils.BoolTo[int](choice))
+	s.V.CMove(x0s.V, x1s.V, utils.BoolTo[uint64](choice))
 	return s
 }
